refactor(spec): separate multi-depth MTP drafter from eagle
Inkling-style MTP is the original multi-token prediction: distinct depth
layers, each re-run over shifted verify windows with per-depth KV repair
(extend catch-up, decode windows, lookback stashes). This is essentially
different from eagle-like MTP (a single layer chained on its own hidden,
e.g. DeepSeek), so host it in its own drafter, like dflash.py:

- Move all multi-depth window/lookback/catch-up logic to a standalone
  Mtp(BaseDrafter) in drafter/mtp.py, dropping the eagle-only residue it
  can never reach (DSA top-k plumbing, EAGLE3 hot-token vocab mapping,
  the legacy singular mm-pad setter). The classic chain loop stays as
  the INKLING_MTP_CLASSIC_LOOP A/B fallback.
- Restore drafter/eagle.py to its pre-Inkling state (93ecb3ae~1),
  keeping only the updated dict-by-modality mm-pad substitution API that
  ModelExecutor now calls.
- Route "MTP" per draft model class in _get_drafter_impl: Inkling's
  NextN model gets Mtp; everything else stays on Eagle.
- Retarget the window/stash helper imports in
  test_drafter_accept_indexing.py to drafter.mtp.

Verified: test_drafter_accept_indexing (13 passed),
test_deepseek_v4_config (107 passed), restored eagle.py diffs against
93ecb3ae~1 only in the mm-pad API. Live-engine coverage is blocked by a
pre-existing test_inkling_mtp_e2e failure at HEAD (FA4 decode kernel
asserts headdim % 64 == 0 against the tiny fixture's head_dim=32).
